refactor: build mode
