+++ package.json
@@ -73,6 +73,6 @@
   },
   "packageManager": "npm@9.8.1",
   "engines": {
-    "node": ">= v16.20.1 || >= v18.18.0 || >= 20.7.0"
+    "node": ">= v16.20.1 || >= v18.18.0 || >= v20.7.0"
   }
 }
